Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/service-workers/stub-3.1-service-worker-obj-expected.txt

CONSOLE MESSAGE: line 871: Invalid object member name ServiceWorkerState

Harness Error (FAIL), message = Invalid object member name ServiceWorkerState

FAIL ServiceWorker interface: existence and properties of interface object assert_equals: prototype of ServiceWorker is not Worker expected function "function Worker() {
    [native code]
}" but got function "function EventTarget() {
    [native code]
}"
PASS ServiceWorker interface object length 
PASS ServiceWorker interface object name 
FAIL ServiceWorker interface: existence and properties of interface prototype object assert_equals: prototype of ServiceWorker.prototype is not Worker.prototype expected object "[object WorkerPrototype]" but got object "[object EventTargetPrototype]"
PASS ServiceWorker interface: existence and properties of interface prototype object's "constructor" property 
PASS ServiceWorker interface: existence and properties of interface prototype object's @@unscopables property 
FAIL ServiceWorker interface: attribute scope assert_true: The prototype object must have a property "scope" expected true got false
FAIL ServiceWorker interface: attribute url assert_true: The prototype object must have a property "url" expected true got false
PASS ServiceWorker interface: attribute state 
PASS ServiceWorker interface: attribute onstatechange 
FAIL ServiceWorker must be primary interface of throw new Error ('No object defined for the ServiceWorker interface') assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorker interface"
FAIL Stringification of throw new Error ('No object defined for the ServiceWorker interface') assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorker interface"
FAIL ServiceWorker interface: throw new Error ('No object defined for the ServiceWorker interface') must inherit property "scope" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorker interface"
FAIL ServiceWorker interface: throw new Error ('No object defined for the ServiceWorker interface') must inherit property "url" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorker interface"
FAIL ServiceWorker interface: throw new Error ('No object defined for the ServiceWorker interface') must inherit property "state" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorker interface"
FAIL ServiceWorker interface: throw new Error ('No object defined for the ServiceWorker interface') must inherit property "onstatechange" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorker interface"
